vgtron will now accept only one (1) target file.
This is mostly for sanity on the part of the implementation of the
helper script, but it has an impact on usage as well.

For implementation, this means I don't have to create another loop
around the loop invoking `vegeta` and incrementing the rate value.

The usage case comes in with the --prefix flag being added shortly. If
you can pass multiple test files, the datafiles produced by `vegeta`
become less meaningful or confusing. One test at a time, one data file.