Output file naming. Command line parsing.
This is a big commit, because most of these features are linked and all
flow into each other.

I was calling parse_args all sorts of wrong. The build_parser() function
wasn't returning the parser it created, flow control was all wrong. Had
to restructure this to make it work whatsoever, and also for
testability.

Output file naming was redesigned a bit. UUIDs are completely excessive.
Too long and convey no meaningful information about the context of the
test which is the *entire point of vgtron*. Files are now named with an
optional prefix and then `r{request_rate}-{timestamp}`.

This provides quite a lot of useful info at-at-glance. Timestamp first
lets you separate the test outputs by when they were started, whcih in
the design of this tool is the 'suite' of tests being run. The rate
allows you to quickly and programmatically determine the request rate.
Everything you need in one concise format.

One small tweak to the range() invocation. Added the initial rate
parameter to the max parameter because otherwise, the max rate isn't
correctly reached.