Tests are no longer run in parallel.
Each Popen() invocation is now chained into a wait() invocation. This
means that each iteration of the loadtest will complete fully before
beginning the next rate-level. This way, if you set a max rate of one
hundred thousand and a step of 100, you won't end up with all ~10,000
iterations of the load test running simultaneously. This kills the
server.